+++ engines/solar2d/config.lua
@@ -1,0 +1,8 @@
+application = {
+  content = {
+    width = 720,
+    height = 1280,
+    scale = "letterbox",
+    fps = 60
+  }
+}